# docs/api-contracts/admin-module-openapi.yaml
...
openapi: 3.0.3
info:
  title: Admin Module API
  description: |
    Admin Module — управляющий модуль системы Artstore.

    **Admin Module не является auth-сервером.** Аутентификация и выдача JWT
    полностью делегированы внешнему Identity Provider (Keycloak). Admin Module
    получает JWT от API Gateway, проверяет claims (роли, scopes) и принимает
    решения по авторизации для своих endpoints.

    ## Аутентификация

    Все запросы (кроме Health) проходят через API Gateway, который:
    - Проверяет подпись JWT через JWKS Keycloak
    - Проверяет expiration
    - Форвардит claims в заголовках к Admin Module

    ### Admin Users (H2M)
    Аутентификация через Keycloak OIDC (Authorization Code + PKCE).
    JWT содержит `sub`, `realm_access.roles`, `groups`.

    Роли определяются маппингом Keycloak groups → Artstore roles:
    - `artstore-admins` → `admin` (полный доступ)
    - `artstore-viewers` → `readonly` (только чтение)

    Admin Module может **дополнить** (но не понизить) роль через
    локальные role overrides.

    ### Service Accounts (M2M)
    OAuth 2.0 Client Credentials flow через Keycloak.
    JWT содержит `client_id`, `scope`.

    Scopes:
    - `files:read` — чтение метаданных и скачивание файлов
    - `files:write` — загрузка, обновление и удаление файлов
    - `storage:read` — чтение информации о Storage Elements
    - `storage:write` — управление SE (sync, mode transition)
    - `admin:read` — чтение административных данных
...

    ## Shared PostgreSQL
...

...

...

...
servers:
...
paths:
...
  /api/v1/service-accounts:
    post:
      tags: [service-accounts]
      summary: Создать сервисный аккаунт
      description: |
        Создание нового SA. Автоматически синхронизируется в Keycloak
        (создаётся confidential client с Client Credentials flow).

        Credentials (client_id, client_secret) генерируются Keycloak.
        **client_secret возвращается только один раз** в ответе на создание.

        Доступно только роли `admin`.
      operationId: createServiceAccount
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceAccountCreate"
      responses:
        "201":
          description: SA создан (client_secret в ответе!)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceAccountWithSecret"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Имя уже занято
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: CONFLICT
                  message: "Сервисный аккаунт с таким именем уже существует"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      tags: [service-accounts]
      summary: Список сервисных аккаунтов
      description: |
        Возвращает пагинированный список SA.
        Доступно ролям `admin` и `readonly`.
      operationId: listServiceAccounts
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: status
          in: query
          description: Фильтр по статусу
          schema:
            type: string
            enum: [active, suspended]
      responses:
        "200":
          description: Список SA
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceAccountListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/service-accounts/{id}:
    get:
      tags: [service-accounts]
      summary: Получить сервисный аккаунт
      description: |
        Возвращает данные SA по ID.
        Доступно ролям `admin` и `readonly`.
      operationId: getServiceAccount
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ServiceAccountId"
      responses:
        "200":
          description: Данные SA
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceAccount"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      tags: [service-accounts]
      summary: Обновить сервисный аккаунт
      description: |
        Обновление SA (name, description, scopes, status).
        Изменения синхронизируются в Keycloak.
        Для изменения secret используйте `rotate-secret`.

        Доступно только роли `admin`.
      operationId: updateServiceAccount
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ServiceAccountId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceAccountUpdate"
      responses:
        "200":
          description: SA обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceAccount"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [service-accounts]
      summary: Удалить сервисный аккаунт
      description: |
        Удаление SA. Синхронно удаляется client в Keycloak.
        Доступно только роли `admin`.
      operationId: deleteServiceAccount
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ServiceAccountId"
      responses:
        "204":
          description: SA удалён
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/service-accounts/{id}/rotate-secret:
    post:
      tags: [service-accounts]
      summary: Ротация secret
      description: |
        Генерирует новый `client_secret` для SA в Keycloak.
        Старый secret немедленно становится невалидным.
        Новый secret возвращается **только один раз** в ответе.

        Доступно только роли `admin`.
      operationId: rotateSecret
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ServiceAccountId"
      responses:
        "200":
          description: Secret обновлён (новый secret в ответе!)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RotateSecretResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # =========================================================================
  # Storage Elements (7 endpoints)
  # =========================================================================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT от Keycloak. Валидация подписи — на API Gateway.
        Admin Module проверяет claims (роли, scopes).

        Для Admin Users: `realm_access.roles`, `groups`
        Для Service Accounts: `scope`

  # =========================================================================
  # Параметры
  # =========================================================================
  parameters:
    Limit:
      name: limit
      in: query
      description: Количество записей на странице
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
    Offset:
      name: offset
      in: query
      description: Смещение от начала списка
      schema:
        type: integer
        minimum: 0
        default: 0
...
    ServiceAccountId:
      name: id
      in: path
      required: true
      description: UUID сервисного аккаунта
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Отсутствует или невалидный JWT token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: "Требуется аутентификация"
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: FORBIDDEN
              message: "Недостаточно прав для выполнения операции"
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: NOT_FOUND
              message: "Ресурс не найден"
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INTERNAL_ERROR
              message: "Внутренняя ошибка сервера"

  # =========================================================================
  # Схемы данных
  # =========================================================================
  schemas:
...
    ServiceAccount:
      type: object
      description: Сервисный аккаунт (без secret)
      required:
        - id
        - client_id
        - name
        - scopes
        - status
        - source
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          example: 660e8400-e29b-41d4-a716-446655440001
        keycloak_client_id:
          type: string
          nullable: true
          description: UUID client в Keycloak
        client_id:
          type: string
          description: Идентификатор для аутентификации (формат `sa_<name>_<random>`)
          example: sa_ingester_a1b2c3
        name:
          type: string
          description: Человекочитаемое имя
          example: ingester
        description:
          type: string
          nullable: true
          example: Сервисный аккаунт для Ingester Module
        scopes:
          type: array
          items:
            type: string
            enum:
              - files:read
              - files:write
              - storage:read
              - storage:write
              - admin:read
              - admin:write
          example: [files:write, storage:read]
        status:
          type: string
          enum: [active, suspended]
          example: active
        source:
          type: string
          enum: [keycloak, local]
          description: |
            Источник создания SA:
            - `local` — создан через Admin Module
            - `keycloak` — обнаружен при синхронизации с Keycloak
          example: local
        last_synced_at:
          type: string
          format: date-time
          nullable: true
          description: Время последней синхронизации с Keycloak
        created_at:
          type: string
          format: date-time
          example: "2026-02-21T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-02-21T10:00:00Z"
    ServiceAccountWithSecret:
      description: SA с secret (возвращается только при создании)
      allOf:
        - $ref: "#/components/schemas/ServiceAccount"
        - type: object
          required:
            - client_secret
          properties:
            client_secret:
              type: string
              description: |
                Секрет для аутентификации от Keycloak.
                **Показывается только один раз!**
              example: cs_random_secret_value
    ServiceAccountCreate:
      type: object
      required:
        - name
        - scopes
      properties:
        name:
          type: string
          description: Имя SA (используется в client_id)
          minLength: 2
          maxLength: 50
          example: ingester
        description:
          type: string
          maxLength: 500
          example: Сервисный аккаунт для Ingester Module
        scopes:
          type: array
          items:
            type: string
            enum:
              - files:read
              - files:write
              - storage:read
              - storage:write
              - admin:read
              - admin:write
          minItems: 1
          example: [files:write, storage:read]
    ServiceAccountUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
        description:
          type: string
          maxLength: 500
        scopes:
          type: array
          items:
            type: string
            enum:
              - files:read
              - files:write
              - storage:read
              - storage:write
              - admin:read
              - admin:write
          minItems: 1
        status:
          type: string
          enum: [active, suspended]
      minProperties: 1
    RotateSecretResponse:
      type: object
      required:
        - client_id
        - client_secret
      properties:
        client_id:
          type: string
          example: sa_ingester_a1b2c3
        client_secret:
          type: string
          description: Новый секрет от Keycloak. **Показывается только один раз!**
          example: cs_new_random_secret
    ServiceAccountListResponse:
      type: object
      required:
        - items
        - total
        - limit
        - offset
        - has_more
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ServiceAccount"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean

    # -----------------------------------------------------------------------
    # Storage Elements
    # -----------------------------------------------------------------------
...
    ErrorResponse:
      type: object
      description: Стандартный формат ошибки (единый для всей системы Artstore)
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: |
                Машиночитаемый код ошибки:
                - `VALIDATION_ERROR` — некорректные входные данные
                - `NOT_FOUND` — ресурс не найден
                - `UNAUTHORIZED` — требуется аутентификация
                - `FORBIDDEN` — недостаточно прав
                - `CONFLICT` — конфликт (дублирующийся ресурс)
                - `SE_UNAVAILABLE` — SE недоступен
                - `IDP_UNAVAILABLE` — Keycloak недоступен
                - `INTERNAL_ERROR` — внутренняя ошибка
              example: NOT_FOUND
            message:
              type: string
              description: Человекочитаемое описание
              example: Ресурс не найден